+++ src/main/java/overlook_hotel/overlook_hotel/repository/ManagerRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import overlook_hotel.overlook_hotel.model.entity.Manager;
 
 @Repository
-public interface ManagerRepository extends JpaRepository<Manager, Integer>, JpaSpecificationExecutor<Manager> {
+public interface ManagerRepository extends JpaRepository<Manager, Long>, JpaSpecificationExecutor<Manager> {
 }
